+++ src/components/Profile/Profile.jsx
@@ -1,0 +1,8 @@
+export const Profile = () => {
+
+  return (
+    <div className="profile">
+      <img src="https://portal.staralliance.com/imagelibrary/aux-pictures/prototype-images/avatar-default.png/@@images/image.png"></img>
+    </div>
+  )
+}
